+++ test/test-helper.js
@@ -45,5 +45,58 @@ module.exports = {
     });
 
     return component;
-  }
+  },
+
+  unmountComponent: function(component) {
+    if(component && component.isMounted()) {
+      React.unmountComponentAtNode(component.getDOMNode().parentNode);
+    }
+  },
+
+  createNativeClickEvent: function() {
+    var evt = document.createEvent('HTMLEvents');
+    evt.initEvent('click', false, true);
+
+    return evt;
+  },
+
+  createNativeMouseEvent: function(options) {
+    var evt = document.createEvent('MouseEvents');
+    evt.initEvent(options.action, false, true);
+
+    return evt;
+  },
+
+  createNativeKeyboardEvent: function(options) {
+    var evt = document.createEvent('HTMLEvents');
+    var keyEvent = options.event || 'keyup';
+    evt.which = options.which;
+    evt.keycode = options.which;
+    evt.initEvent(keyEvent, false, true);
+
+    return evt;
+  },
+
+  noop: function() {},
+
+  keyCodes: {
+    BACKSPACE: 8,
+    TAB: 9,
+    ENTER: 13,
+    ESCAPE: 27,
+    SPACE: 32,
+    PAGE_UP: 33,
+    PAGE_DOWN: 34,
+    END: 35,
+    HOME: 36,
+    LEFT: 37,
+    UP: 38,
+    RIGHT: 39,
+    DOWN: 40,
+    DELETE: 46,
+    COMMA: 188,
+    SHIFT: 16,
+    CTRL: 17,
+    ALT: 18
+}
 };