+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Change Log
 
-## master
+## 0.3.0
 
+- updated testing system
+- added ability to specify specific file for the copy static assets gulp task
+- made browserify error easier to work with
 - refactored some test code to play nicely with [react-router](https://github.com/rackt/react-router)
 - converted to using [react-router](https://github.com/rackt/react-router)
 - fixed issue with browserify-production not uglifying code